# src/main/resources/swagger/openapi-doc.yml
openapi: 3.0.1

info:
  title: Kavaleridze Museum API
  description: |
    The Kavaleridze Museum API is a comprehensive RESTful API that serves as the backbone of the Kavaleridze Museum's digital platform. It provides a wide range of endpoints for managing museum events, images, articles, email communications, and search functionalities. This API empowers administrators, visitors, and staff to interact with museum-related content, enhancing the overall visitor experience and facilitating efficient museum management.
  version: v0.0.1

tags:
  - name: Admin Event Controller
    description: |
      Endpoints for administrators to manage museum events. Includes creating, updating, deleting events, and more.
    externalDocs:
      url: /admin/events
  - name: Admin Images Controller
    description: |
      Endpoints for administrators to manage images on the website. Provides operations for uploading, deleting, and organizing images.
    externalDocs:
      url: /admin/images
  - name: Event Controller
    description: |
      Endpoints for retrieving museum events. Allows users to view event details, search for events, and more.
    externalDocs:
      url: /events
  - name: Image Controller
    description: |
      Endpoints to retrieve and display images from the storage. Provides access to images for use on the website.
    externalDocs:
      url: /images
  - name: Mail Controller
    description: |
      Endpoints for email communications, including sending feedback and inquiries. Enables communication with museum staff.
    externalDocs:
      url: /feedback
  - name: Search Controller
    description: |
      Endpoints for performing searches on the website. Allows users to search for events, articles, and more.
    externalDocs:
      url: /search
  - name: Article Controller
    description: |
      Endpoints for performing article-related operations. Provides access to articles and allows users to interact with them.
    externalDocs:
      url: /articles

servers:
  - url: http://127.0.0.1/api
    description: Service for managing website of the Kavaleridze museum

paths:
  /admin/events:
    post:
      tags:
        - Admin Event Controller
      summary: Create a new Event
      description: Creates a new Event with the provided information.
      operationId: createEvent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventPublication'
        required: true
      responses:
        201:
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/events/{id}:
    put:
      tags:
        - Admin Event Controller
      summary: Update an existing Event
      description: Updates an existing Event with the provided information.
      operationId: updateEvent
      parameters:
        - name: id
          in: path
          description: Unique identifier of the Event to be updated
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventPublication'
        required: true
      responses:
        200:
          description: Event updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Admin Event Controller
      summary: Delete an existing Event
      description: Deletes an existing Event by its unique identifier.
      operationId: deleteEvent
      parameters:
        - name: id
          in: path
          description: Unique identifier of the Event to be deleted
          required: true
          schema:
            type: string
      responses:
        204:
          description: Event deleted successfully
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/images:
    post:
      tags:
        - Admin Images Controller
      operationId: saveImage
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                type: string
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/images/temp:
    get:
      tags:
        - Admin Images Controller
      operationId: getTempImage
      parameters:
        - name: filename
          in: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            image/jpeg:
              schema:
                type: string
                format: byte
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /images:
    get:
      tags:
        - Image Controller
      operationId: getImage
      parameters:
        - name: filename
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            image/jpeg:
              schema:
                type: string
                format: byte
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /feedback/submit:
    post:
      tags:
        - Mail Controller
      operationId: submitContactForm
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MailDto'
        required: true
      responses:
        200:
          description: Email sent successfully
          content:
            application/json:
              schema:
                type: object
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /events:
    get:
      tags:
        - Event Controller
      summary: Get a page of events
      description: Retrieves a page of events with the specified size and page number.
      operationId: getAll
      parameters:
        - name: size
          in: query
          description: Size of the page
          required: true
          schema:
            type: integer
            format: int32
        - name: page
          in: query
          description: Page number
          required: true
          schema:
            type: integer
            format: int32
      responses:
        200:
          description: Page of events retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageEvent'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /events/{id}:
    get:
      tags:
        - Event Controller
      summary: Get an Event by ID
      description: Retrieves an Event by its unique identifier.
      operationId: getById
      parameters:
        - name: id
          in: path
          description: Unique identifier of the Event
          required: true
          schema:
            type: string
      responses:
        200:
          description: Event retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /articles/{title}:
    get:
      tags:
        - Article Controller
      summary: Find article by title
      description: Returns an article by its title.
      operationId: findByTitle
      parameters:
        - name: title
          in: path
          description: Title of the article
          required: true
          schema:
            type: string
      responses:
        200:
          description: Found the article
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        400:
          description: Title request is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Article not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /search/{query}:
    get:
      tags:
        - Search Controller
      summary: Perform a full text search search
      description: Performs a full text search search in the website content based on the provided query string.
      operationId: search
      parameters:
        - name: query
          in: path
          description: Search query string
          required: true
          schema:
            minLength: 3
            maxLength: 120
            type: string
      responses:
        200:
          description: Search results retrieved successfully or an empty payload
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchResponse'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        timestamp:
          type: integer
          format: int64
    EventPublication:
      required:
        - description
        - title
        - type
      type: object
      properties:
        title:
          type: string
          nullable: false
        description:
          type: string
          nullable: false
        type:
          type: string
          nullable: false
        tags:
          uniqueItems: true
          type: array
          items:
            type: string
        content:
          type: string
          minLength: 30
          maxLength: 3000
          nullable: false
        bannerTempURI:
          type: string
        begin:
          type: string
          format: date
          nullable: false
        end:
          type: string
          format: date
          nullable: false
      description: Event data to be created
    EventResponse:
      required:
        - begin
        - description
        - end
        - title
      type: object
      properties:
        id:
          type: string
        title:
          type: string
          minLength: 3
          maxLength: 300
        description:
          type: string
          minLength: 3
          maxLength: 300
        type:
          type: string
        tags:
          type: array
          uniqueItems: true
          items:
            type: string
        content:
          type: string
          minLength: 30
          maxLength: 3000
        bannerURI:
          type: string
        bannerPreviewURI:
          type: string
        begin:
          type: string
          format: date
        end:
          type: string
          format: date
        created:
          type: string
          format: date
        updated:
          type: string
          format: date
    MailDto:
      required:
        - email
        - firstName
        - lastName
        - message
      type: object
      properties:
        firstName:
          type: string
          minLength: 2
          maxLength: 30
          nullable: false
        lastName:
          type: string
          minLength: 2
          maxLength: 30
          nullable: false
        email:
          type: string
          format: email
          nullable: false
        message:
          type: string
          minLength: 10
          maxLength: 300
          nullable: false
      description: Feedback form data
    Article:
      required:
        - content
        - description
        - title
      type: object
      properties:
        id:
          type: string
        title:
          type: string
          minLength: 3
          maxLength: 300
        description:
          type: string
          minLength: 3
          maxLength: 300
        content:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ContentBlock'
        created:
          type: string
          format: date
        updated:
          type: string
          format: date
    ContentBlock:
      required:
        - blockType
        - columns
        - order
      type: object
      properties:
        id:
          type: string
        order:
          type: integer
          format: int32
        columns:
          type: integer
          format: int32
        blockType:
          type: string
          enum:
            - TEXT_BLOCK
            - PICTURE_BLOCK
            - PICTURE_TEXT_BLOCK
        textContent:
          type: string
        pictureLink:
          type: string
    PageEvent:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        first:
          type: boolean
        pageable:
          $ref: '#/components/schemas/PageableObject'
        last:
          type: boolean
        sort:
          $ref: '#/components/schemas/SortObject'
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/EventResponse'
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
        sort:
          $ref: '#/components/schemas/SortObject'
        offset:
          type: integer
          format: int64
    SortObject:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
    SearchResponse:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        contentType:
          type: string
          enum:
            - ARTICLE
            - EVENT